8단계 구현완료

+++ src/main/java/com/example/todolist/Controller/ScheduleController.java
@@ -4,6 +4,8 @@ import com.example.todolist.Dto.Schedule.ScheduleListResponse;
 import com.example.todolist.Dto.Schedule.ScheduleRequestDto;
 import com.example.todolist.Dto.Schedule.ScheduleResponseDto;
 import com.example.todolist.Entity.Schedule;
+import com.example.todolist.Exception.InvalidPasswordException;
+import com.example.todolist.Exception.ScheduleNotFoundException;
 import com.example.todolist.Repository.Schedule.ScheduleRepository;
 import com.example.todolist.Service.Schedule.ScheduleServiceImpl;
 import lombok.RequiredArgsConstructor;
@@ -33,7 +35,7 @@ public class ScheduleController {
     @GetMapping("/{sid}")
     public ResponseEntity<ScheduleResponseDto> getSchedule(@PathVariable Long sid) {
         if(scheduleService.readByScheduleId(sid) == null ){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            throw new ScheduleNotFoundException("일정이 존재하지 않습니다.");
         }
         return ResponseEntity.ok(scheduleService.readByScheduleId(sid));
     }
@@ -50,7 +52,7 @@ public class ScheduleController {
     @PutMapping("/update/{sid}")
     public ResponseEntity<Schedule> updateSchedule(@PathVariable Long sid, @RequestBody ScheduleRequestDto scheduleRequestDto) {
         if(scheduleRequestDto.getPassword() == null || scheduleRequestDto.getPassword().isEmpty()){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            throw new InvalidPasswordException("비밀번호가 일치 하지 않습니다.");
         }
              Schedule update = scheduleService.updateTaskAndManager_id(sid,scheduleRequestDto);
              return ResponseEntity.ok(update);
@@ -59,8 +61,11 @@ public class ScheduleController {
 
     @DeleteMapping("/delete/{sid}")
     public ResponseEntity<Schedule> deleteSchedule(@PathVariable Long sid) {
-        Schedule schedule= scheduleService.deleteBySid(sid);
-        return new ResponseEntity<>(schedule,HttpStatus.OK);
+        if(scheduleService.readByScheduleId(sid) == null){
+            throw new ScheduleNotFoundException("이미 삭제된 일정입니다.");
+        }
+        scheduleService.deleteBySid(sid);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
 }
